+++ fe/fe-core/src/test/java/org/apache/doris/qe/HmsQueryCacheTest.java
@@ -51,8 +51,6 @@ import org.junit.jupiter.api.Test;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.locks.ReentrantReadWriteLock;
-
 
 public class HmsQueryCacheTest extends AnalyzeCheckTestBase {
     private static final String HMS_CATALOG = "hms_ctl";
@@ -119,7 +117,6 @@ public class HmsQueryCacheTest extends AnalyzeCheckTestBase {
         Deencapsulation.setField(db, "initialized", true);
 
         Deencapsulation.setField(tbl, "objectCreated", true);
-        Deencapsulation.setField(tbl, "rwLock", new ReentrantReadWriteLock(true));
         Deencapsulation.setField(tbl, "schemaUpdateTime", NOW);
         Deencapsulation.setField(tbl, "eventUpdateTime", 0);
         new Expectations(tbl) {
@@ -167,7 +164,6 @@ public class HmsQueryCacheTest extends AnalyzeCheckTestBase {
         };
 
         Deencapsulation.setField(tbl2, "objectCreated", true);
-        Deencapsulation.setField(tbl2, "rwLock", new ReentrantReadWriteLock(true));
         Deencapsulation.setField(tbl2, "schemaUpdateTime", NOW);
         Deencapsulation.setField(tbl2, "eventUpdateTime", 0);
         new Expectations(tbl2) {
@@ -215,7 +211,6 @@ public class HmsQueryCacheTest extends AnalyzeCheckTestBase {
         };
 
         Deencapsulation.setField(view1, "objectCreated", true);
-        Deencapsulation.setField(view1, "rwLock", new ReentrantReadWriteLock(true));
 
         new Expectations(view1) {
             {
@@ -270,7 +265,6 @@ public class HmsQueryCacheTest extends AnalyzeCheckTestBase {
         };
 
         Deencapsulation.setField(view2, "objectCreated", true);
-        Deencapsulation.setField(view2, "rwLock", new ReentrantReadWriteLock(true));
         new Expectations(view2) {
             {
                 view2.getId();
